xattr: allow reading of xattrs
This adds support for reading extended attributes from disk.
Attributes() returning a fs.AttributesInfo is added to fs.Entry. Getting
the attributes from disk uses github.com/pkg/xattr.

fs.AttributesInfo is a map, which when empty is just a pointer. It being
a map (as opposed to a slice in restic) helps in comparing if the
attributes have changed (no need to allocate a map there and then).

Most other changes in snapshot/snapshotfs/* are there to implement the
extended fs.Entry interface.

Signed-off-by: Miek Gieben <miek@miek.nl>