Update _index.md
Add context to why `--insecure` cannot be used behind a reverse proxy.

I am using kopia behind traefik as a reverse proxy. With the current wording, I thought that the lack of --insecure was required by the nginx (config) and not kopia itself. [This thread](https://kopia.discourse.group/t/cant-connect-to-insecure-repository-server/871/4) finally helped me to resolve the connection issues to the repo server.