Simplify SetCounters signature
Pass arguments in a `restore.Stats` struct.
  `SetCounters(s restore.Stats)`

Simplifies call sites and implementation.

In this case it makes sense to pass all the values
using the restore.Stats struct as it simplifies
the calls.

However, this pattern should be avoided in general
as it essentially makes all the arguments "optional".
This makes it easy to miss setting a value and simply
passing 0 (the default value), thus it becomes error
prone.
In this particular case, the struct is being passed
through verbatim, thus eliminiting the risk of
missing a value, at least in the current state of
the code.