fix(snapshots): stop passing down fuse Context
a fuse-provided Context is canceled when the originating process of the
request receives a signal before the request is answered; in such cases
we should return EINTR instead of EIO to not fail the request entirely.
since each context is associated to a single request this requires some
kind of rebinding of existing fuse structures to new contexts, which we
do not have at the moment, we can only ignore the fuse request context.

fixes #4383