Use omitzero in `ActionsPolicy`
The `omitempty` and `omitzero` have practically the same effect on
fields of pointer types:
- the field is omitted when it is null
- the field is included when it is not null, even if the value
  that it points to is "the zero value" for the non-pointer type.

Note: when the pointer type defines an `IsZero` member function,
then that field would also be omitted during marshaling.
There is no defined `IsZero` function for these pointers, so
the sematics are preserved in this case.

The `omitzero` JSON tag in the fields in `ActionPolicyDefinition`
does not change the semantics, it simply makes the marshaled
representation more compact.